+++ bgstally/activity.py
@@ -1,10 +1,10 @@
 import json
 import re
 from copy import deepcopy
-from datetime import datetime, timedelta
+from datetime import UTC, datetime, timedelta
 from typing import Dict
 
-from bgstally.constants import FILE_SUFFIX, CheckStates
+from bgstally.constants import ApiSizeLookup, ApiSyntheticEvent, ApiSyntheticCZObjectiveType, ApiSyntheticScenarioType, DATETIME_FORMAT_JOURNAL, FILE_SUFFIX, CheckStates
 from bgstally.debug import Debug
 from bgstally.missionlog import MissionLog
 from bgstally.state import State
@@ -89,7 +89,7 @@ SPACECZ_PILOTNAMES_SPECOPS = [
     '$LUASC_Scenario_Warzone_NPC_SpecOps_D;'
 ]
 
-SPACECZ_PILOTNAME_PROPAGAND = '$LUASC_Scenario_Warzone_NPC_WarzoneCorrespondent;'
+SPACECZ_PILOTNAME_CORRESPONDENT = '$LUASC_Scenario_Warzone_NPC_WarzoneCorrespondent;'
 
 CZ_GROUND_LOW_CB_MAX = 5000
 CZ_GROUND_MED_CB_MAX = 38000
@@ -126,13 +126,14 @@ class Activity:
     factions with their activity
     """
 
-    def __init__(self, bgstally, tick: Tick = None, sample: bool = False):
+    def __init__(self, bgstally, tick: Tick = None, sample: bool = False, cmdr = None):
         """Constructor
 
         Args:
             bgstally (BGSTally): The BGSTally object
             tick (Tick, optional): The Tick object to instantiate from. If None, the last known tick is used. Defaults to None.
             sample (bool, optional): Populate with sample data. Defaults to False.
+            cmdr (str, optional): The CMDR name. This is not done properly (yet) - the cmdr name is simply updated often to be the latest cmdr seen.
         """
         self.bgstally = bgstally
         if tick == None: tick = Tick(self.bgstally)
@@ -144,6 +145,8 @@ class Activity:
         self.discord_webhook_data:dict = {} # key = webhook uuid, value = dict containing webhook data
         self.discord_notes: str = ""
         self.dirty: bool = False
+
+        self.cmdr: str = cmdr  # Not saved / loaded (yet) because it's not implemented properly
 
         if sample:
             self.systems: dict = {"Sample System ID": self.get_sample_system_data()}
@@ -557,23 +560,27 @@ class Activity:
             self.recalculate_zero_activity()
 
 
-    def bv_received(self, journal_entry: Dict, state: State):
+    def bv_received(self, journal_entry: Dict, state: State, cmdr: str):
+        """Handle a bounty voucher for a kill
+
+        Args:
+            journal_entry (Dict): The journal data
+            state (State): The bgstally State object
+            cmdr (str): The CMDR name
         """
-        Handle a bounty voucher for a kill
-        """
-        current_system = self.systems.get(state.current_system_id)
+        current_system: dict = self.systems.get(state.current_system_id)
         if not current_system: return
 
         # Check whether in megaship scenario for scenario tracking
         if state.last_megaship_approached != {}:
-            timedifference = datetime.strptime(journal_entry['timestamp'], "%Y-%m-%dT%H:%M:%SZ") - datetime.strptime(state.last_megaship_approached['timestamp'], "%Y-%m-%dT%H:%M:%SZ")
+            timedifference: datetime = datetime.strptime(journal_entry['timestamp'], DATETIME_FORMAT_JOURNAL) - datetime.strptime(state.last_megaship_approached['timestamp'], DATETIME_FORMAT_JOURNAL)
             if timedifference > timedelta(minutes=5):
                 # Too long since we last entered a megaship scenario, we can't be sure we're fighting at that scenario, clear down
                 state.last_megaship_approached = {}
             else:
                 # We're within the timeout, refresh timestamp and handle the CB
                 state.last_megaship_approached['timestamp'] = journal_entry['timestamp']
-                self._bv_megaship_scenario(journal_entry, current_system, state)
+                self._bv_megaship_scenario(journal_entry, current_system, state, cmdr)
 
 
     def bv_redeemed(self, journal_entry: Dict, state: State):
@@ -596,9 +603,13 @@ class Activity:
                 self.recalculate_zero_activity()
 
 
-    def cb_received(self, journal_entry: dict, state: State):
-        """
-        Handle a combat bond received for a kill
+    def cb_received(self, journal_entry: dict, state: State, cmdr: str):
+        """Handle a combat bond received for a kill
+
+        Args:
+            journal_entry (dict): The journal data
+            state (State): The bgstally State object
+            cmdr (str): The CMDR name
         """
         current_system = self.systems.get(state.current_system_id)
         if not current_system: return
@@ -610,7 +621,7 @@ class Activity:
 
         # Otherwise, must be on-ground or in-space CZ for CB kill tracking
         if state.last_settlement_approached != {}:
-            timedifference = datetime.strptime(journal_entry['timestamp'], "%Y-%m-%dT%H:%M:%SZ") - datetime.strptime(state.last_settlement_approached['timestamp'], "%Y-%m-%dT%H:%M:%SZ")
+            timedifference = datetime.strptime(journal_entry['timestamp'], DATETIME_FORMAT_JOURNAL) - datetime.strptime(state.last_settlement_approached['timestamp'], DATETIME_FORMAT_JOURNAL)
             if timedifference > timedelta(minutes=5):
                 # Too long since we last approached a settlement, we can't be sure we're fighting at that settlement, clear down
                 state.last_settlement_approached = {}
@@ -618,17 +629,17 @@ class Activity:
             else:
                 # We're within the timeout, refresh timestamp and handle the CB
                 state.last_settlement_approached['timestamp'] = journal_entry['timestamp']
-                self._cb_ground_cz(journal_entry, current_system, state)
+                self._cb_ground_cz(journal_entry, current_system, state, cmdr)
 
         elif state.last_spacecz_approached != {}:
-            timedifference = datetime.strptime(journal_entry['timestamp'], "%Y-%m-%dT%H:%M:%SZ") - datetime.strptime(state.last_spacecz_approached['timestamp'], "%Y-%m-%dT%H:%M:%SZ")
+            timedifference = datetime.strptime(journal_entry['timestamp'], DATETIME_FORMAT_JOURNAL) - datetime.strptime(state.last_spacecz_approached['timestamp'], DATETIME_FORMAT_JOURNAL)
             if timedifference > timedelta(minutes=5):
                 # Too long since we last entered a space cz, we can't be sure we're fighting at that cz, clear down
                 state.last_spacecz_approached = {}
             else:
                 # We're within the timeout, refresh timestamp and handle the CB
                 state.last_spacecz_approached['timestamp'] = journal_entry['timestamp']
-                self._cb_space_cz(journal_entry, current_system, state)
+                self._cb_space_cz(journal_entry, current_system, state, cmdr)
 
 
     def cb_redeemed(self, journal_entry: Dict, state: State):
@@ -647,7 +658,7 @@ class Activity:
             self.recalculate_zero_activity()
 
 
-    def cap_ship_bond_received(self, journal_entry: dict):
+    def cap_ship_bond_received(self, journal_entry: dict, cmdr: str):
         """Handle a capital ship bond
 
         Args:
@@ -663,6 +674,14 @@ class Activity:
             self.dirty = True
 
             faction['SpaceCZ']['cs'] = int(faction['SpaceCZ'].get('cs', '0')) + 1
+
+            event: dict = {
+                'event': ApiSyntheticEvent.CZOBJECTIVE,
+                'count': 1,
+                'type': ApiSyntheticCZObjectiveType.CAPSHIP,
+                'Faction': faction
+            }
+            self.bgstally.api_manager.send_event(event, self, cmdr)
 
             self.bgstally.ui.show_system_report(current_system['SystemAddress'])
             self.recalculate_zero_activity()
@@ -973,11 +992,17 @@ class Activity:
         self.bgstally.ui.show_system_report(current_system['SystemAddress'])
 
 
-    def _cb_ground_cz(self, journal_entry:dict, current_system:dict, state:State):
+    def _cb_ground_cz(self, journal_entry: dict, current_system: dict, state: State, cmdr: str):
+        """Combat bond received while we are in an active ground CZ
+
+        Args:
+            journal_entry (dict): The journal entry data
+            current_system (dict): The current system data
+            state (State): The bgstally State object
+            cmdr (str): The CMDR name
         """
-        Combat bond received while we are in an active ground CZ
-        """
-        faction = current_system['Factions'].get(journal_entry['AwardingFaction'])
+        faction_name: str = journal_entry.get('AwardingFaction', "")
+        faction: dict = current_system['Factions'].get(faction_name)
         if not faction: return
 
         self.dirty = True
@@ -989,7 +1014,7 @@ class Activity:
             faction['GroundCZSettlements'][state.last_settlement_approached['name']] = self._get_new_groundcz_settlement_data()
 
         # Store the previously counted size of this settlement
-        previous_size = state.last_settlement_approached['size']
+        previous_size: str = state.last_settlement_approached['size']
 
         # Increment this settlement's overall count if this is the first bond counted
         if state.last_settlement_approached['size'] == None:
@@ -1007,6 +1032,16 @@ class Activity:
                 faction['GroundCZSettlements'][state.last_settlement_approached['name']]['type'] = 'l'
                 # Store last settlement type
                 state.last_settlement_approached['size'] = 'l'
+
+                # Send to API
+                event: dict = {
+                    'event': ApiSyntheticEvent.GROUNDCZ,
+                    'low': 1,
+                    'settlement': state.last_settlement_approached['name'],
+                    'Faction': faction_name
+                }
+                self.bgstally.api_manager.send_event(event, self, cmdr)
+
         elif journal_entry['Reward'] < CZ_GROUND_MED_CB_MAX:
             # Handle as 'Med' if this is either the first CB or we've counted this settlement as a 'Low' before
             if state.last_settlement_approached['size'] == None or state.last_settlement_approached['size'] == 'l':
@@ -1018,6 +1053,16 @@ class Activity:
                 faction['GroundCZSettlements'][state.last_settlement_approached['name']]['type'] = 'm'
                 # Store last settlement type
                 state.last_settlement_approached['size'] = 'm'
+
+                # Send to API
+                event: dict = {
+                    'event': ApiSyntheticEvent.GROUNDCZ,
+                    'medium': 1,
+                    'settlement': state.last_settlement_approached['name'],
+                    'Faction': faction_name
+                }
+                if previous_size != None: event[ApiSizeLookup[previous_size]] = -1
+                self.bgstally.api_manager.send_event(event, self, cmdr)
         else:
             # Handle as 'High' if this is either the first CB or we've counted this settlement as a 'Low' or 'Med' before
             if state.last_settlement_approached['size'] == None or state.last_settlement_approached['size'] == 'l' or state.last_settlement_approached['size'] == 'm':
@@ -1030,19 +1075,31 @@ class Activity:
                 # Store last settlement type
                 state.last_settlement_approached['size'] = 'h'
 
+                # Send to API
+                event: dict = {
+                    'event': ApiSyntheticEvent.GROUNDCZ,
+                    'high': 1,
+                    'settlement': state.last_settlement_approached['name'],
+                    'Faction': faction_name
+                }
+                if previous_size != None: event[ApiSizeLookup[previous_size]] = -1
+                self.bgstally.api_manager.send_event(event, self, cmdr)
+
         self.recalculate_zero_activity()
 
 
-    def _cb_space_cz(self, journal_entry:dict, current_system:dict, state:State):
+    def _cb_space_cz(self, journal_entry: dict, current_system: dict, state: State, cmdr: str):
         """Combat bond received while we are in an active space CZ
 
         Args:
             journal_entry (dict): The journal entry data
-            current_system (dict): The current system dict
+            current_system (dict): The current system data
             state (State): The bgstally state object
+            cmdr (str): The CMDR name
         """
 
-        faction = current_system['Factions'].get(journal_entry.get('AwardingFaction', ""))
+        faction_name: str = journal_entry.get('AwardingFaction', "")
+        faction: dict = current_system['Factions'].get(faction_name)
         if not faction: return
 
         # Check for side objectives detected by CBs
@@ -1051,18 +1108,48 @@ class Activity:
                 # Tally a captain kill. Unreliable because of journal order unpredictability.
                 state.last_spacecz_approached['capt'] = True
                 faction['SpaceCZ']['cp'] = int(faction['SpaceCZ'].get('cp', '0')) + 1
+
+                # Send to API
+                event: dict = {
+                    'event': ApiSyntheticEvent.CZOBJECTIVE,
+                    'count': 1,
+                    'type': ApiSyntheticCZObjectiveType.GENERAL,
+                    'Faction': faction_name
+                }
+                self.bgstally.api_manager.send_event(event, self, cmdr)
+
                 self.bgstally.ui.show_system_report(current_system['SystemAddress'])
             elif state.last_ship_targeted.get('PilotName', "") in SPACECZ_PILOTNAMES_SPECOPS and not state.last_spacecz_approached.get('specops'):
                 # Tally a specops kill. We would like to only tally this after 4 kills in a CZ, but sadly due to journal order
                 # unpredictability we tally as soon as we spot a kill after targeting a spec ops
                 state.last_spacecz_approached['specops'] = True
                 faction['SpaceCZ']['so'] = int(faction['SpaceCZ'].get('so', '0')) + 1
+
+                # Send to API
+                event: dict = {
+                    'event': ApiSyntheticEvent.CZOBJECTIVE,
+                    'count': 1,
+                    'type': ApiSyntheticCZObjectiveType.SPECOPS,
+                    'Faction': faction_name
+                }
+                self.bgstally.api_manager.send_event(event, self, cmdr)
+
                 self.bgstally.ui.show_system_report(current_system['SystemAddress'])
-            elif state.last_ship_targeted.get('PilotName', "") == SPACECZ_PILOTNAME_PROPAGAND and not state.last_spacecz_approached.get('propagand'):
+            elif state.last_ship_targeted.get('PilotName', "") == SPACECZ_PILOTNAME_CORRESPONDENT and not state.last_spacecz_approached.get('propagand'):
                 # Tally a propagandist kill. We would like to only tally this after 3 kills in a CZ, but sadly due to journal order
                 # unpredictability we tally as soon as we spot a kill after targeting a propagandist
                 state.last_spacecz_approached['propagand'] = True
                 faction['SpaceCZ']['pr'] = int(faction['SpaceCZ'].get('pr', '0')) + 1
+
+                # Send to API
+                event: dict = {
+                    'event': ApiSyntheticEvent.CZOBJECTIVE,
+                    'count': 1,
+                    'type': ApiSyntheticCZObjectiveType.CORRESPONDENT,
+                    'Faction': faction_name
+                }
+                self.bgstally.api_manager.send_event(event, self, cmdr)
+
                 self.bgstally.ui.show_system_report(current_system['SystemAddress'])
 
         # If we've already counted this CZ, exit
@@ -1072,20 +1159,35 @@ class Activity:
         state.last_spacecz_approached['ally_faction'] = faction.get('Faction', "")
         self.dirty = True
 
-        type:str = state.last_spacecz_approached.get('type', 'l')
+        type: str = state.last_spacecz_approached.get('type', 'l')
         faction['SpaceCZ'][type] = int(faction['SpaceCZ'].get(type, '0')) + 1
+
+        # Send to API
+        event: dict = {
+            'event': ApiSyntheticEvent.CZ,
+            ApiSizeLookup[type]: 1,
+            'Faction': faction_name
+        }
+        self.bgstally.api_manager.send_event(event, self, cmdr)
 
         self.bgstally.ui.show_system_report(current_system['SystemAddress'])
         self.recalculate_zero_activity()
 
 
-    def _bv_megaship_scenario(self, journal_entry:dict, current_system:dict, state:State):
+    def _bv_megaship_scenario(self, journal_entry: dict, current_system: dict, state: State, cmdr: str):
+        """We are in an active megaship scenario
+
+        Args:
+            journal_entry (dict): The journal entry data
+            current_system (dict): The current system data
+            state (State): The bgstally State object
+            cmdr (str): The CMDR name
         """
-        We are in an active megaship scenario
-        """
-        faction:dict = current_system['Factions'].get(journal_entry['VictimFaction'])
+        faction_name: str = journal_entry.get('VictimFaction', "")
+        faction: dict = current_system['Factions'].get(faction_name)
         if not faction: return
-        opponent_faction:dict = current_system['Factions'].get(faction.get('Opponent', ""))
+        opponent_faction_name: str = faction.get('Opponent', "")
+        opponent_faction: dict = current_system['Factions'].get(opponent_faction_name)
         if not opponent_faction: return
 
         # If we've already counted this scenario, exit
@@ -1096,6 +1198,15 @@ class Activity:
 
         # The scenario should be counted against the opponent faction of the ship just killed
         opponent_faction['Scenarios'] += 1
+
+        # Send to API
+        event: dict = {
+            'event': ApiSyntheticEvent.SCENARIO,
+            'type': ApiSyntheticScenarioType.MEGASHIP,
+            'count': 1,
+            'Faction': opponent_faction_name
+        }
+        self.bgstally.api_manager.send_event(event, self, cmdr)
 
         self.bgstally.ui.show_system_report(current_system['SystemAddress'])
         self.recalculate_zero_activity()
@@ -1368,6 +1479,7 @@ class Activity:
         """
         self.tick_id = dict.get('tickid')
         self.tick_time = datetime.strptime(dict.get('ticktime'), DATETIME_FORMAT_ACTIVITY)
+        self.tick_time = self.tick_time.replace(tzinfo=UTC)
         self.tick_forced = dict.get('tickforced', False)
         self.discord_webhook_data = dict.get('discordwebhookdata', {})
         self.discord_notes = dict.get('discordnotes', "")